Extract task package into struct in separate module
This commit refactors the way we handle task packages in the SDK.
It renames `Location` to `Package`, and adds a method on top `publish`
which for `Package::Archive(PathBuf)` is responsible for publishing
the archive over `gftp` and computing the archive's digest; both
ops are async.

This is also the first step at general refactoring of the SDK's structure.